+++ prisma/migrations/20250815093913_name_index_in_company/migration.sql
@@ -1,0 +1,2 @@
+-- CreateIndex
+CREATE INDEX "Company_name_idx" ON "public"."Company"("name");